+++ payments/RELEASES.md
@@ -228,3 +228,13 @@ Based on:
 - [typescript v2.8.2] payments
 ### Releases
 - [NPM v2.8.2] https://www.npmjs.com/package/@wingspan/payments/v/2.8.2 - payments
+
+## 2024-12-02 01:32:54
+### Changes
+Based on:
+- OpenAPI Doc 1.0.0-oas3.1 ./openapi/payment_speakeasy_edit.yaml
+- Speakeasy CLI 1.448.2 (2.467.4) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v2.8.3] payments
+### Releases
+- [NPM v2.8.3] https://www.npmjs.com/package/@wingspan/payments/v/2.8.3 - payments